Fix uint64 typed array gen

+++ testing/cbor_map_gen.go
@@ -73,7 +73,7 @@ func (t *SimpleTypeTree) MarshalCBOR(w io.Writer) error {
 		return err
 	}
 	for _, v := range t.Others {
-		if err := cbg.CborWriteHeader(w, cbg.MajUnsignedInt, v); err != nil {
+		if err := cbg.CborWriteHeader(w, cbg.MajUnsignedInt, uint64(v)); err != nil {
 			return err
 		}
 	}
@@ -238,7 +238,7 @@ func (t *SimpleTypeTree) UnmarshalCBOR(r io.Reader) error {
 					return xerrors.Errorf("value read for array t.Others was not a uint, instead got %d", maj)
 				}
 
-				t.Others[i] = val
+				t.Others[i] = uint64(val)
 			}
 
 			// t.Test ([][]uint8) (slice)
